moves column range into row filter module
