+++ lib/row_filter/row_filter.ex
@@ -1,6 +1,7 @@
 defmodule Bigtable.RowFilter do
-  alias Bigtable.RowFilter.ColumnRange
-  alias Google.Bigtable.V2.{ReadRowsRequest, RowFilter, TimestampRange}
+  alias Google.Bigtable.V2.{ColumnRange, ReadRowsRequest, RowFilter, TimestampRange}
+
+  @type column_range :: {binary(), binary(), boolean()} | {binary(), binary()}
 
   @moduledoc """
   Provides functions for creating `Google.Bigtable.V2.RowFilter` and applying them to a `Google.Bigtable.V2.ReadRowsRequest` or `Google.Bigtable.V2.RowFilter.Chain`.
@@ -330,7 +331,7 @@ defmodule Bigtable.RowFilter do
   @spec column_range(binary(), {binary(), binary(), boolean()} | {binary(), binary()}) ::
           RowFilter.t()
   def column_range(family_name, range) do
-    range = ColumnRange.create_range(family_name, range)
+    range = create_range(family_name, range)
 
     {:column_range_filter, range}
     |> build_filter()
@@ -545,5 +546,42 @@ defmodule Bigtable.RowFilter do
   @spec apply_filter(RowFilter.t(), ReadRowsRequest.t()) :: ReadRowsRequest.t()
   defp apply_filter(%RowFilter{} = filter, %ReadRowsRequest{} = request) do
     %{request | filter: filter}
+  end
+
+  @spec create_range(binary(), column_range) :: ColumnRange.t()
+  def create_range(family_name, {start_qualifier, end_qualifier, inclusive}) do
+    range = translate_range(start_qualifier, end_qualifier, inclusive)
+
+    range
+    |> Keyword.put(:family_name, family_name)
+    |> ColumnRange.new()
+  end
+
+  def create_range(family_name, {start_qualifier, end_qualifier}) do
+    create_range(family_name, {start_qualifier, end_qualifier, true})
+  end
+
+  @spec translate_range(binary(), binary(), boolean()) :: Keyword.t()
+  defp translate_range(start_qualifier, end_qualifier, inclusive) do
+    case inclusive do
+      true -> inclusive_range(start_qualifier, end_qualifier)
+      false -> exclusive_range(start_qualifier, end_qualifier)
+    end
+  end
+
+  @spec exclusive_range(binary(), binary()) :: Keyword.t()
+  defp exclusive_range(start_qualifier, end_qualifier) do
+    [
+      start_qualifier: {:start_qualifier_open, start_qualifier},
+      end_qualifier: {:end_qualifier_open, end_qualifier}
+    ]
+  end
+
+  @spec inclusive_range(binary(), binary()) :: Keyword.t()
+  defp inclusive_range(start_qualifier, end_qualifier) do
+    [
+      start_qualifier: {:start_qualifier_closed, start_qualifier},
+      end_qualifier: {:end_qualifier_closed, end_qualifier}
+    ]
   end
 end
